+++ src/test/java/com/tsp/new_tsp_admin/api/model/service/negotiation/AdminNegotiationJpaRepositoryTest.java
@@ -171,7 +171,7 @@ class AdminNegotiationJpaRepositoryTest {
                 .modelNegotiationDesc("영화 프로젝트 참여 테스트 두번째").modelNegotiationDate(now()).build());
 
         List<AdminModelDTO> modelNegotiationList = new ArrayList<>();
-        modelNegotiationList.add(AdminModelDTO.builder().idx(3).categoryCd(1).modelKorName("조찬희")
+        modelNegotiationList.add(AdminModelDTO.builder().idx(3L).categoryCd(1).modelKorName("조찬희")
                 .modelNegotiation(negotiationList).build());
 
         // when
@@ -206,7 +206,7 @@ class AdminNegotiationJpaRepositoryTest {
                 .modelNegotiationDesc("영화 프로젝트 참여 테스트 두번째").modelNegotiationDate(now()).build());
 
         List<AdminModelDTO> modelNegotiationList = new ArrayList<>();
-        modelNegotiationList.add(AdminModelDTO.builder().idx(3).categoryCd(1).modelKorName("조찬희")
+        modelNegotiationList.add(AdminModelDTO.builder().idx(3L).categoryCd(1).modelKorName("조찬희")
                 .modelNegotiation(negotiationList).build());
 
         // when
@@ -227,7 +227,7 @@ class AdminNegotiationJpaRepositoryTest {
     @DisplayName("이전 or 다음 모델 섭외 상세 조회 테스트")
     void 이전or다음모델섭외상세조회테스트() {
         // given
-        adminNegotiationEntity = AdminNegotiationEntity.builder().idx(2).build();
+        adminNegotiationEntity = AdminNegotiationEntity.builder().idx(2L).build();
 
         // when
         adminNegotiationDTO = adminNegotiationJpaRepository.findOneNegotiation(adminNegotiationEntity);
@@ -242,7 +242,7 @@ class AdminNegotiationJpaRepositoryTest {
     @DisplayName("이전 모델 섭외 상세 조회 Mockito 테스트")
     void 이전모델섭외상세조회Mockito테스트() {
         // given
-        adminNegotiationEntity = AdminNegotiationEntity.builder().idx(2).build();
+        adminNegotiationEntity = AdminNegotiationEntity.builder().idx(2L).build();
 
         // when
         adminNegotiationDTO = adminNegotiationJpaRepository.findOneNegotiation(adminNegotiationEntity);
@@ -266,7 +266,7 @@ class AdminNegotiationJpaRepositoryTest {
     @DisplayName("이전 모델 섭외 상세 조회 BDD 테스트")
     void 이전모델섭외상세조회BDD테스트() {
         // given
-        adminNegotiationEntity = AdminNegotiationEntity.builder().idx(2).build();
+        adminNegotiationEntity = AdminNegotiationEntity.builder().idx(2L).build();
 
         // when
         adminNegotiationDTO = adminNegotiationJpaRepository.findOneNegotiation(adminNegotiationEntity);
@@ -287,7 +287,7 @@ class AdminNegotiationJpaRepositoryTest {
     @DisplayName("다음 모델 섭외 상세 조회 Mockito 테스트")
     void 다음모델섭외상세조회Mockito테스트() {
         // given
-        adminNegotiationEntity = AdminNegotiationEntity.builder().idx(2).build();
+        adminNegotiationEntity = AdminNegotiationEntity.builder().idx(2L).build();
 
         // when
         adminNegotiationDTO = adminNegotiationJpaRepository.findOneNegotiation(adminNegotiationEntity);
@@ -311,7 +311,7 @@ class AdminNegotiationJpaRepositoryTest {
     @DisplayName("다음 모델 섭외 상세 조회 BDD 테스트")
     void 다음모델섭외상세조회BDD테스트() {
         // given
-        adminNegotiationEntity = AdminNegotiationEntity.builder().idx(2).build();
+        adminNegotiationEntity = AdminNegotiationEntity.builder().idx(2L).build();
 
         // when
         adminNegotiationDTO = adminNegotiationJpaRepository.findOneNegotiation(adminNegotiationEntity);
@@ -377,7 +377,7 @@ class AdminNegotiationJpaRepositoryTest {
     @DisplayName("모델섭외수정Mockito테스트")
     void 모델섭외수정Mockito테스트() {
         // given
-        Integer idx = adminNegotiationJpaRepository.insertModelNegotiation(adminNegotiationEntity).getIdx();
+        Long idx = adminNegotiationJpaRepository.insertModelNegotiation(adminNegotiationEntity).getIdx();
 
         adminNegotiationEntity = AdminNegotiationEntity.builder()
                 .idx(idx)
@@ -416,7 +416,7 @@ class AdminNegotiationJpaRepositoryTest {
     @DisplayName("모델섭외수정BDD테스트")
     void 모델섭외수정BDD테스트() {
         // given
-        Integer idx = adminNegotiationJpaRepository.insertModelNegotiation(adminNegotiationEntity).getIdx();
+        Long idx = adminNegotiationJpaRepository.insertModelNegotiation(adminNegotiationEntity).getIdx();
 
         adminNegotiationEntity = AdminNegotiationEntity.builder()
                 .idx(idx)
@@ -457,7 +457,7 @@ class AdminNegotiationJpaRepositoryTest {
 
         // when
         when(mockAdminNegotiationJpaRepository.findOneNegotiation(adminNegotiationEntity)).thenReturn(adminNegotiationDTO);
-        Integer deleteIdx = adminNegotiationJpaRepository.deleteModelNegotiation(adminNegotiationEntity.getIdx());
+        Long deleteIdx = adminNegotiationJpaRepository.deleteModelNegotiation(adminNegotiationEntity.getIdx());
 
         // then
         assertThat(mockAdminNegotiationJpaRepository.findOneNegotiation(adminNegotiationEntity).getIdx()).isEqualTo(deleteIdx);
@@ -480,7 +480,7 @@ class AdminNegotiationJpaRepositoryTest {
 
         // when
         given(mockAdminNegotiationJpaRepository.findOneNegotiation(adminNegotiationEntity)).willReturn(adminNegotiationDTO);
-        Integer deleteIdx = adminNegotiationJpaRepository.deleteModelNegotiation(adminNegotiationEntity.getIdx());
+        Long deleteIdx = adminNegotiationJpaRepository.deleteModelNegotiation(adminNegotiationEntity.getIdx());
 
         // then
         assertThat(mockAdminNegotiationJpaRepository.findOneNegotiation(adminNegotiationEntity).getIdx()).isEqualTo(deleteIdx);
